+++ 20220122_DavidKwon7/presentation/src/main/java/com/example/presentation/FavoriteActivity.kt
@@ -72,7 +72,7 @@ class FavoriteActivity : AppCompatActivity() {
     private fun initVMFactory(){
         val viewModelFactory = ViewModelProvider(
             this,ViewModelProvider.NewInstanceFactory())
-            .get(MainViewModel::class.java)
+            .get(FavoriteViewModel::class.java)
     }
